N9i

+++ philo/ft_sleep.c
@@ -6,26 +6,26 @@
 /*   By: ozahidi <ozahidi@student.1337.ma>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/20 11:39:49 by ozahidi           #+#    #+#             */
-/*   Updated: 2024/05/20 12:10:02 by ozahidi          ###   ########.fr       */
+/*   Updated: 2024/05/23 17:33:08 by ozahidi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-size_t get_current_time(void)
+long	get_current_time(void)
 {
-    struct timeval tv;
+	struct timeval	tv;
 
-    gettimeofday(&tv, NULL);
-    return ((tv.tv_sec * 1000) + (tv.tv_usec / 1000));
+	gettimeofday(&tv, NULL);
+	return ((tv.tv_sec * 1000) + (tv.tv_usec / 1000));
 }
 
-int ft_sleep(int time)
+int	ft_sleep(int time)
 {
-    size_t start;
+	long	start;
 
-    start = get_current_time();
-    while (get_current_time() - start < time)
-        usleep(100);
-    return (1);
+	start = get_current_time();
+	while (get_current_time() - start < time)
+		usleep(100);
+	return (1);
 }
